requirements.txt
gTTS == 2.1.1
opencv-python == 4.2.0.32
Pillow == 7.0.0
pydub == 0.23.1
SpeechRecognition == 3.8.1
